+++ src/configuration/models/uniteTheme/uniteThemeConfiguration.ts
@@ -2,6 +2,7 @@
  * Model of Unite Theme Configuration (unite-theme.json) file.
  */
 export class UniteThemeConfiguration {
+    public shortName: string;
     public metaAuthor: string;
     public metaDescription: string;
     public metaKeywords: string[];
@@ -9,4 +10,6 @@ export class UniteThemeConfiguration {
     public themeHeaders: string[];
     public backgroundColor: string;
     public themeColor: string;
+    public appLoaderStyle: string;
+    public appLoader: string;
 }